GS: Use 16-bit indices instead of 32-bit
Save some bandwidth.

+++ pcsx2/GS/GSState.h
@@ -158,7 +158,7 @@ protected:
 
 	struct
 	{
-		u32* buff;
+		u16* buff;
 		u32 tail;
 	} m_index = {};
